boards/nrf52840dk: split into lib+bin crates for configuration/tutorial boards
